Kernel: Move AC'97 to its own subdirectory
