+++ Kernel/CMakeLists.txt
@@ -42,7 +42,7 @@ set(KERNEL_SOURCES
     Bus/VirtIO/Queue.cpp
     Bus/VirtIO/RNG.cpp
     Devices/AsyncDeviceRequest.cpp
-    Devices/Audio/AC97.cpp
+    Devices/Audio/AC97/AC97.cpp
     Devices/Audio/Channel.cpp
     Devices/Audio/IntelHDA/Codec.cpp
     Devices/Audio/IntelHDA/Controller.cpp
